app/requirements.txt
fastapi==0.111.0 # Latest stable version
uvicorn==0.30.1 # Latest stable version
google-adk==0.0.10
python-dotenv==1.0.1